# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Users criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar Users",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar todos Users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/roles": {
      "get": {
        "operationId": "UsersController_getRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation getRoles",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Users",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Users",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Users",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/warranties": {
      "post": {
        "operationId": "WarrantiesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Warranties criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar Warranties",
        "tags": [
          "Warranties"
        ]
      },
      "get": {
        "operationId": "WarrantiesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar todos Warranties",
        "tags": [
          "Warranties"
        ]
      }
    },
    "/warranties/{id}": {
      "get": {
        "operationId": "WarrantiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Warranties",
        "tags": [
          "Warranties"
        ]
      },
      "delete": {
        "operationId": "WarrantiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Warranties",
        "tags": [
          "Warranties"
        ]
      }
    },
    "/warranties/{id}/status": {
      "patch": {
        "operationId": "WarrantiesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation updateStatus",
        "tags": [
          "Warranties"
        ]
      }
    },
...
  },
  "info": {
    "title": "Click Marido API",
    "description": "API documentation for Click Marido ERP/CRM",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Campo email",
            "example": "exemplo"
          },
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "password": {
            "type": "string",
            "description": "Campo password",
            "example": "exemplo"
          },
          "roleIds": {
            "description": "Campo roleIds",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Campo isActive",
            "example": true
          }
        },
        "required": [
          "email",
          "name",
          "password",
          "roleIds"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Campo email",
            "example": "exemplo"
          },
          "name": {
            "type": "string",
            "description": "Campo name",
            "example": "exemplo"
          },
          "password": {
            "type": "string",
            "description": "Campo password",
            "example": "exemplo"
          },
          "roleIds": {
            "description": "Campo roleIds",
            "example": "exemplo",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Campo isActive",
            "example": true
          }
        }
      },
...
    }
  }
}